Safari-609.1.11.1/LayoutTests/fast/events/before-input-prevent-undo-expected.txt

To manually test this, type some into the input field below and try to undo. The value of the text field should not change after attempting to undo.

initial value:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "abc"

after prevented undo:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "abc"
